Update event.ps1
`event.ps1` will now continue listening for the length entered in the code. Previously it would hear one change from the `TEMP` directory, copy the virus, then just continue the countdown without continuing to listen for changes in `TEMP`. If the virus is deleted from the main directory it was copied to, `event.ps1` will initiate another copy of the virus to replace it.